Adding JPA annotations to Entities

+++ back-end/bug-tracker-ws/src/main/java/com/adanh/ws/repository/BugRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.adanh.ws.entity.Bug;
 
 @Repository
-public interface BugRepository extends JpaRepository<Bug, Long> {
+public interface BugRepository extends JpaRepository<Bug, Integer> {
 	
 
 }
