+++ back-end/bug-tracker-ws/src/main/java/com/adanh/ws/repository/DocumentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.adanh.ws.entity.Document;
 
 @Repository
-public interface DocumentRepository extends JpaRepository<Document, Long> {
+public interface DocumentRepository extends JpaRepository<Document, Integer> {
 	
 
 }
